src/requirements.txt
annotated-types==0.7.0
asciitree==0.3.3
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
czifile==2019.7.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dicomweb-client==0.59.3
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
fasteners==0.19
fsspec==2024.12.0
gunicorn==23.0.0
idna==3.10
imagecodecs==2024.12.30
marshmallow==3.26.1
numcodecs==0.15.1
numpy==2.2.0
ome-types==0.5.3
openslide-bin==4.0.0.6
openslide-python==1.4.1
opentile==0.15.0
packaging==24.2
pillow==10.4.0
pycparser==2.22
pydantic==2.10.6
pydantic-compat==0.1.2
pydantic_core==2.27.2
pydicom==3.0.1
pyisyntax==0.1.4
pyjpegls==1.5.1
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.3.0
pylibjpeg-openjpeg==2.4.0
pylibjpeg-rle==2.0.0
PySyntax==0.0.0.dev6
python-dotenv==1.0.1
PyTurboJPEG==1.7.7
requests==2.32.3
retrying==1.3.4
six==1.17.0
sqlparse==0.5.3
tifffile==2025.2.18
tiffslide==2.5.0
typing_extensions==4.12.2
universal_pathlib==0.2.6
urllib3==2.3.0
wrapt==1.17.2
wsidicom==0.24.0
wsidicomizer==0.18.0
xsdata==24.3.1
zarr==2.18.4
